+++ modules/gametree.js
@@ -72,24 +72,16 @@ context.getRoot = function(tree) {
 context.navigate = function(tree, index, step) {
     if (index + step >= 0 && index + step < tree.nodes.length) {
         return [tree, index + step]
-    } else if (index + step < 0 && tree.parent) {
-        if (tree.parent != null) {
-            var prev = tree.parent
-            var newstep = index + step + 1
+    } else if (index + step < 0 && tree.parent != null) {
+        var prev = tree.parent
+        var newstep = index + step + 1
 
-            return context.navigate(prev, prev.nodes.length - 1, newstep)
-        }
+        return context.navigate(prev, prev.nodes.length - 1, newstep)
+    } else if (index + step >= tree.nodes.length && tree.current != null) {
+        var next = tree.subtrees[tree.current]
+        var newstep = index + step - tree.nodes.length
 
-        return [tree, 0]
-    } else if (index + step >= tree.nodes.length) {
-        if (tree.current != null) {
-            var next = tree.subtrees[tree.current]
-            var newstep = index + step - tree.nodes.length
-
-            return context.navigate(next, 0, newstep)
-        }
-
-        return [tree, tree.nodes.length - 1]
+        return context.navigate(next, 0, newstep)
     }
 
     return [null, 0]
@@ -284,7 +276,7 @@ context.matrixdict2graph = function(matrixdict) {
                 data: matrix[y][x],
                 originalColor: setting.get('graph.node_color')
             }
-            var commentproperties = ['C', 'UC', 'GW', 'DM', 'GB', 'BM', 'TE', 'DO', 'IT']
+            var commentproperties = ['C', 'N', 'UC', 'GW', 'DM', 'GB', 'BM', 'TE', 'DO', 'IT']
 
             if (commentproperties.some(function(x) { return x in tree.nodes[index] }))
                 node.originalColor = setting.get('graph.node_comment_color')